SUPER MEGA COMPLETE COOL ULTRA OVERHAUL!!
I made a visualization handler which is the entity that handles all the visualizations
like a nice zookeeper!

We can now dynamically create and delete visualizations.
Information is persistent between visualizations, click on one and this will be shared to the others. (and color the correct cell)
You can select a visualization by clicking on it, after this you can drag and zoom on it.

Also improved performance just to add it to the commit.