Made more progress towards the canvas-based visualization handler
Instead of the visualization handler having an 'active' visualization it now has a dictionary, which I had to create from scratch, that maps a canvas to a visualization. Using this mapping I deprecated every function that worked with the old 'active' principe and changed it all to work with the 'selected' principe, where every function expect the canvas that called the function as an argument and thus used to this to handle the call.

I will in a later commit remove the deprecated functions, but not yet.