Moved functions from the canvas to the abstract class.
This way people designing visualizations can change and or disable functions that had implemented in the canvas.
In this case I am talking about the zooming and the dragging

Most of the calculations of the zooming and dragging have been moved to the abstract class under similar names.
The only thing remaining is the decision making if something was a drag or not.